requirements.txt
# Database Connection (SQL Server)
pyodbc==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Machine Learning (Your Jupyter Logic)
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4

# Visualization (Your Plots)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
kaleido==0.2.1

# Desktop GUI (Interactive Interface)
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1

# Report Generation (Excel/PDF Export)
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
Pillow==10.1.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Optional - Performance
numba==0.58.1
